RF 04.00 | Folder > Create folder using invalid data
